Update alerts-overview.md
We need to rectify this document where it says in Automation we should use external unique ID because we don't get any details of external unique ID while make Graph API call to trigger Playbook on alerts also this is legacy way. Here we need to update that DetectorID should be used which never changes and you get the details of Detector through Microsoft graph api calls.